+++ settings.ini
@@ -1,0 +1,10 @@
+[settings]
+scanner_window = (0, 216, 512, 576)
+image_scale = 0.685
+confidence_threshold = 0.9
+display_inventory_items = False
+display_unavailable_recipes = False
+copy_recipe_to_clipboard = False
+scan_hotkey = 
+run_as_overlay = True
+
