+++ src/UIOverlay.py
@@ -1,262 +1,16 @@
-import sys
-from dataclasses import dataclass
 from configparser import ConfigParser
-
-import win32gui
-from win32clipboard import *
-
-import tkinter as tk
-from tkinter import messagebox
-from typing import Callable, Any, Tuple, List, Dict
-
-import cv2
-import numpy as np
-from PIL import ImageTk, Image, ImageGrab
-
 import keyboard
+import tkinter as tk
+import sys
+from win32clipboard import OpenClipboard, EmptyClipboard, SetClipboardText, CloseClipboard
 
-COLOR_BG = 'grey19'
-COLOR_FG_WHITE = 'snow'
-COLOR_FG_GREEN = 'green3'
-COLOR_FG_LIGHT_GREEN = 'DarkOliveGreen3'
-COLOR_FG_ORANGE = 'orange2'
-FONT_BIG = ('Consolas', '14')
-FONT_SMALL = ('Consolas', '9')
+from typing import Dict, List, Tuple
+from DataClasses import PoeWindowInfo
+from ArchnemesisItemsMap import ArchnemesisItemsMap
+from ImageScanner import ImageScanner
+from DataClasses import RecipeItemNode
+from constants import COLOR_BG, COLOR_FG_GREEN, COLOR_FG_LIGHT_GREEN, COLOR_FG_ORANGE, COLOR_FG_WHITE, FONT_BIG, FONT_SMALL
 
-@dataclass
-class RecipeItemNode:
-    item: str
-    components: list
-
-class ArchnemesisItemsMap:
-    """
-    Holds the information about all archnemesis items, recipes, images and map them together
-    """
-    def __init__(self, scale: float):
-        # Put everything into the list so we could maintain the display order
-        self._arch_items = [
-            ('Kitava-Touched', ['Tukohama-Touched', 'Abberath-Touched', 'Corrupter', 'Corpse Detonator']),
-            ('Innocence-Touched', ['Lunaris-Touched', 'Solaris-Touched', 'Mirror Image', 'Mana Siphoner']),
-            ('Shakari-Touched', ['Entangler', 'Soul Eater', 'Drought Bringer']),
-            ('Abberath-Touched', ['Flame Strider', 'Frenzied', 'Rejuvenating']),
-            ('Tukohama-Touched', ['Bonebreaker', 'Executioner', 'Magma Barrier']),
-            ('Brine King-Touched', ['Ice Prison', 'Storm Strider', 'Heralding Minions']),
-            ('Arakaali-Touched', ['Corpse Detonator', 'Entangler', 'Assassin']),
-            ('Solaris-Touched', ['Invulnerable', 'Magma Barrier', 'Empowered Minions']),
-            ('Lunaris-Touched', ['Invulnerable', 'Frost Strider', 'Empowered Minions']),
-            ('Effigy', ['Hexer', 'Malediction', 'Corrupter']),
-            ('Empowered Elements', ['Evocationist', 'Steel-Infused', 'Chaosweaver']),
-            ('Crystal-Skinned', ['Permafrost', 'Rejuvenating', 'Berserker']),
-            ('Invulnerable', ['Sentinel', 'Juggernaut', 'Consecrator']),
-            ('Corrupter', ['Bloodletter', 'Chaosweaver']),
-            ('Mana Siphoner', ['Consecrator', 'Dynamo']),
-            ('Storm Strider', ['Stormweaver', 'Hasted']),
-            ('Mirror Image', ['Echoist', 'Soul Conduit']),
-            ('Magma Barrier', ['Incendiary', 'Bonebreaker']),
-            ('Evocationist', ['Flameweaver', 'Frostweaver', 'Stormweaver']),
-            ('Corpse Detonator', ['Necromancer', 'Incendiary']),
-            ('Flame Strider', ['Flameweaver', 'Hasted']),
-            ('Soul Eater', ['Soul Conduit', 'Necromancer', 'Gargantuan']),
-            ('Ice Prison', ['Permafrost', 'Sentinel']),
-            ('Frost Strider', ['Frostweaver', 'Hasted']),
-            ('Treant Horder', ['Toxic', 'Sentinel', 'Steel-Infused']),
-            ('Temporal Bubble', ['Juggernaut', 'Hexer', 'Arcane Buffer']),
-            ('Entangler', ['Toxic', 'Bloodletter']),
-            ('Drought Bringer', ['Malediction', 'Deadeye']),
-            ('Hexer', ['Chaosweaver', 'Echoist']),
-            ('Executioner', ['Frenzied', 'Berserker']),
-            ('Rejuvenating', ['Gargantuan', 'Vampiric']),
-            ('Necromancer', ['Bombardier', 'Overcharged']),
-            ('Trickster', ['Overcharged', 'Assassin', 'Echoist']),
-            ('Assassin', ['Deadeye', 'Vampiric']),
-            ('Empowered Minions', ['Necromancer', 'Executioner', 'Gargantuan']),
-            ('Heralding Minions', ['Dynamo', 'Arcane Buffer']),
-            ('Arcane Buffer', []),
-            ('Berserker', []),
-            ('Bloodletter', []),
-            ('Bombardier', []),
-            ('Bonebreaker', []),
-            ('Chaosweaver', []),
-            ('Consecrator', []),
-            ('Deadeye', []),
-            ('Dynamo', []),
-            ('Echoist', []),
-            ('Flameweaver', []),
-            ('Frenzied', []),
-            ('Frostweaver', []),
-            ('Gargantuan', []),
-            ('Hasted', []),
-            ('Incendiary', []),
-            ('Juggernaut', []),
-            ('Malediction', []),
-            ('Opulent', []),
-            ('Overcharged', []),
-            ('Permafrost', []),
-            ('Sentinel', []),
-            ('Soul Conduit', []),
-            ('Steel-Infused', []),
-            ('Stormweaver', []),
-            ('Toxic', []),
-            ('Vampiric', [])
-        ]
-        self._images = dict()
-        self._small_image_size = 30
-        self._update_images(scale)
-
-    def _update_images(self, scale):
-        self._scale = scale
-        for item, _ in self._arch_items:
-            self._images[item] = dict()
-            image = self._load_image(item, scale)
-            self._image_size = image.size
-            self._images[item]['scan-image'] = self._create_scan_image(image)
-            # Convert the image to Tk image because we're going to display it
-            self._images[item]['display-image'] = ImageTk.PhotoImage(image=image)
-            image = image.resize((self._small_image_size, self._small_image_size))
-            self._images[item]['display-small-image'] = ImageTk.PhotoImage(image=image)
-
-    def _load_image(self, item: str, scale: float):
-        image = Image.open(f'pictures/{item}.png')
-        # Scale the image according to the input parameter
-        return image.resize((int(image.width * scale), int(image.height * scale)))
-
-    def _create_scan_image(self, image):
-        # Remove alpha channel and replace it with predefined background color
-        background = Image.new('RGBA', image.size, (10, 10, 32))
-        image_without_alpha = Image.alpha_composite(background, image)
-        scan_template = cv2.cvtColor(np.array(image_without_alpha), cv2.COLOR_RGB2BGR)
-        w, h, _ = scan_template.shape
-
-        # Crop the image to help with scanning
-        return scan_template[int(h * 1.0 / 10):int(h * 2.3 / 3), int(w * 1.0 / 6):int(w * 5.5 / 6)]
-
-
-    def get_scan_image(self, item):
-        return self._images[item]['scan-image']
-
-    def get_display_image(self, item):
-        return self._images[item]['display-image']
-
-    def get_display_small_image(self, item):
-        return self._images[item]['display-small-image']
-
-    def items(self):
-        for item, _ in self._arch_items:
-            yield item
-
-    def recipes(self):
-        for item, recipe in self._arch_items:
-            if recipe:
-                yield (item, recipe)
-
-    def get_subtree_for(self, item: str):
-        tree = RecipeItemNode(item, [])
-        nodes = [tree]
-        while len(nodes) > 0:
-            node = nodes.pop(0)
-            children = self._get_item_components(node.item)
-            if len(children) > 0:
-                node.components = [RecipeItemNode(c, []) for c in children]
-                nodes.extend(node.components)
-        return tree
-
-    def get_parent_recipes_for(self, item: str) -> []:
-        parents = list()
-        for parent, components in self._arch_items:
-            if item in components:
-                parents.append(parent)
-        return parents
-
-    def _get_item_components(self, item) -> List[str]:
-        return next(l for x, l in self._arch_items if x == item)
-
-    @property
-    def image_size(self):
-        return self._image_size
-
-    @property
-    def scale(self) -> float:
-        return self._scale
-
-    @scale.setter
-    def scale(self, scale: float) -> None:
-        self._update_images(scale)
-
-    @property
-    def small_image_size(self):
-        return self._small_image_size
-
-@dataclass
-class PoeWindowInfo:
-    x: int = 0
-    y: int = 0
-    width: int = 0
-    height: int = 0
-    client_width: int = 0
-    client_height: int = 0
-    title_bar_height: int = 0
-
-class ImageScanner:
-    """
-    Implements scanning algorithm with OpenCV. Maintans the scanning window to speed up the scanning.
-    """
-    def __init__(self, info: PoeWindowInfo, items_map: ArchnemesisItemsMap):
-        self._scanner_window_size = (
-            info.x,
-            info.y + int(info.client_height / 4),
-            int(info.client_width / 3),
-            int(info.client_height * 2 / 3)
-        )
-        self._items_map = items_map
-        self._confidence_threshold = 0.94
-
-    def scan(self) -> Dict[str, List[Tuple[int, int]]]:
-        bbox = (
-            self._scanner_window_size[0],
-            self._scanner_window_size[1],
-            self._scanner_window_size[0] + self._scanner_window_size[2],
-            self._scanner_window_size[1] + self._scanner_window_size[3]
-        )
-        screen = ImageGrab.grab(bbox=bbox)
-        screen = np.array(screen)
-        screen = cv2.cvtColor(screen, cv2.COLOR_RGB2BGR)
-
-        results = dict()
-
-        for item in self._items_map.items():
-            template = self._items_map.get_scan_image(item)
-            heat_map = cv2.matchTemplate(screen, template, cv2.TM_CCOEFF_NORMED)
-            _, confidence, _, (x, y) = cv2.minMaxLoc(heat_map)
-            print(f'Best match for {item}: x={x}, y={y} confidence={confidence}', 'too low' if confidence < self._confidence_threshold else '')
-            findings = np.where(heat_map >= self._confidence_threshold)
-            if len(findings[0]) > 0:
-                rectangles = []
-                ht, wt = template.shape[0], template.shape[1]
-                for (x, y) in zip(findings[1], findings[0]):
-                     # Add every box to the list twice in order to retain single (non-overlapping) boxes
-                    rectangles.append([int(x), int(y), int(wt), int(ht)])
-                    rectangles.append([int(x), int(y), int(wt), int(ht)])
-
-                rectangles, _ = cv2.groupRectangles(rectangles, 1, 0.1)
-                results[item] = [(rect[0], rect[1]) for rect in rectangles]
-        print(results)
-        return results
-
-    @property
-    def scanner_window_size(self) -> Tuple[int, int, int, int]:
-        return self._scanner_window_size
-
-    @scanner_window_size.setter
-    def scanner_window_size(self, value: Tuple[int, int, int, int]) -> None:
-        self._scanner_window_size = value
-
-    @property
-    def confidence_threshold(self) -> float:
-        return self._confidence_threshold
-
-    @confidence_threshold.setter
-    def confidence_threshold(self, value) -> None:
-        self._confidence_threshold = value
 
 class UIOverlay:
     """
@@ -508,7 +262,6 @@ class UIOverlay:
     def run(self) -> None:
         self._root.mainloop()
 
-
 class Settings:
     def __init__(self, root, items_map, image_scanner):
         self._root = root
@@ -671,64 +424,3 @@ class Settings:
 
     def get_scan_hotkey(self) -> str:
         return self._scan_hotkey
-
-def show_warning(text: str) -> None:
-    messagebox.showwarning('poe-archnemesis-scanner', text)
-
-def show_error_and_die(text: str) -> None:
-    # Dealing with inconveniences as Perl would
-    messagebox.showerror('poe-archnemesis-scanner', text)
-    sys.exit()
-
-def get_poe_window_info() -> PoeWindowInfo:
-    info = PoeWindowInfo()
-    hwnd = win32gui.FindWindow(None, 'Path of Exile')
-    if hwnd == 0:
-        show_error_and_die('Path of Exile is not running.')
-
-    x0, y0, x1, y1 = win32gui.GetWindowRect(hwnd)
-    info.x = x0
-    info.y = y0
-    info.width = x1 - x0
-    info.height = y1 - y0
-    x0, y0, x1, y1 = win32gui.GetClientRect(hwnd)
-    info.client_width = x1 - x0
-    info.client_height = y1 - y0
-
-    if info.client_width == 0 or info.client_height == 0:
-        show_warning("Unable to detect Path of Exile resolution. Make sure it isn't running in the Fullscreen mode.\n\nThe tool will use your screen resolution for calculations instead.")
-        screen = ImageGrab.grab()
-        info.x = 0
-        info.y = 0
-        info.width, info.height = screen.size
-        info.client_width, info.client_height = screen.size
-    info.title_bar_height = info.height - info.client_height
-    return info
-
-def calculate_default_scale(info: PoeWindowInfo) -> float:
-    """
-    TODO: validate the math for non 16:9 resolutions (e.g. ultrawide monitors)
-    """
-
-    # Assume that all source images have 78x78 size
-    source_image_height = 78.0
-
-    # Take 0.91 as a golden standard for 2560x1440 resolution and calculate
-    # scales for other resolutions based on that
-    constant = 1440.0 / (source_image_height * 0.91)
-    scale = info.client_height / (source_image_height * constant)
-
-    return scale
-
-# Create root as early as possible to initialize some modules (e.g. ImageTk)
-root = tk.Tk()
-root.withdraw()
-
-info = get_poe_window_info()
-
-items_map = ArchnemesisItemsMap(calculate_default_scale(info))
-
-image_scanner = ImageScanner(info, items_map)
-
-overlay = UIOverlay(root, info, items_map, image_scanner)
-overlay.run()
