+++ src/constants.py
@@ -1,0 +1,7 @@
+COLOR_BG = 'grey19'
+COLOR_FG_WHITE = 'snow'
+COLOR_FG_GREEN = 'green3'
+COLOR_FG_LIGHT_GREEN = 'DarkOliveGreen3'
+COLOR_FG_ORANGE = 'orange2'
+FONT_BIG = ('Consolas', '14')
+FONT_SMALL = ('Consolas', '9')
